create @magenta/protobuff, fix core build for npm
export tensorflow.magenta.INoteSequence from protobuff
