+++ magenta/js/core/.gitignore
