+++ magenta/js/core/.npmignore
