+++ magenta/js/core/src/constants.ts
