+++ magenta/js/core/src/sequences.ts
