+++ magenta/js/core/src/sequences_test.ts
