+++ magenta/js/core/tsconfig.jsomn
