+++ magenta/js/core/tslint.json
