+++ magenta/js/protobuff/package.json
@@ -1,0 +1,30 @@
+{
+  "name": "@magenta/protobuff",
+  "version": "0.0.1",
+  "description": "",
+  "main": "src/proto.js",
+  "types": "src/proto.d.ts",
+  "dependencies": {
+    "protobufjs": "^6.8.6"
+	},
+  "devDependencies": {
+    "@types/tape": "^4.2.31",
+    "clang-format": "^1.2.2",
+    "protobufjs": "^6.8.6",
+    "tape": "^4.9.0",
+    "ts-node": "^5.0.0",
+    "tslint": "~5.8.0",
+    "typescript": "2.7.2"
+  },
+  "scripts": {
+    "lint": "tslint -p . -t verbose",
+    "test": "ts-node node_modules/tape/bin/tape src/**/*_test.ts",
+    "proto": "pbjs -t static-module -w commonjs -o src/proto.js ../../protobuf/music.proto && pbts -o src/proto.d.ts src/proto.js"
+  },
+  "author": "Magenta",
+  "license": "Apache-2.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/tensorflow/magenta.git"
+  }
+}
