+++ magenta/js/protobuff/update_proto.sh
